refactor: avoid CLS on gallery page
I have tried leveraging Astro Images better so that it handles the optimization of images better.

It seems that I can set the Image width to essentially control the resolution and then use a surrounding div to control the size of the image on the page, which I do in some cases.